Add SSL to Docker, fix puma, with logger
Its now possible to serve the docker container with an ssl certificate

no more booting rails with rails s.. which defaults to rails s puma.

from now on we start off with puma -e production/developmen

it seems that rails server overwrite the ssl bindings

this was a pain to debug and the only option was to use pure puma server. and add a Rails logger.
so we keep the Active record trace logs.