Improve capistrano's start workers task.
This here adds a prefix 'bundle exec' to the 'rake' command and adds a prefix 'nohup' to it.
All that while preserving the enviroment variables.

Using nohup to silence the rake task. documentation of capistrano on rake task is rather vage.
Especially when dealing with ENV.
Used this as resource instead:

https://github.com/capistrano/sshkit