Improve workers launch in deploy flow.
Workers now stop when the deployment process is in the 'started' fase.
And the start again when it is in the 'finished' fase.

Reference:
http://capistranorb.com/documentation/getting-started/flow/

Also permanently undeamonized sneakers.
Relying on nohup for pushing it to the background.
And collecting the log in rake.out file.
It seems when starting `workers:run` or `sneakers:run`
the make 2 PID while sneakers only sees 1 PID.
Killing both PID is important to ensure complete workers shutdown.