Add a wrapper to parse the response body json before saving
If json is parsed to a hash before saving it to the database as jsonb,
when it is queried it returns as a hash.

I dont know if this is a bug or a feature in rails 5.2.0
because in earlier versions of rails it used to be a hash by default.

I created this wrapper just in case it gets triggerd before_save.
like this im sure of consitency in my database and in my code.

The only think i could find on this online is:
https://github.com/rails/rails/issues/25594