Correct relationships between All Inquiries and Received Inquiry Responses
An Inquirie only need to have 1 Received Inquiry response at any given time,
There is no need for has_many in this case, also no need for the dynamic
interpolation on the relationship level.
Changed the Inquiry relation to has_one Received Inquiry Response.
Changed the Received Inquiry Response to belongs_to Inquiry.

Now comes the interesting part:

Noticed i did not mention Moving, or Cleaning, or Combined.

I call them all Inquiry. but in fact they each and all have their individual name.

The ReceivedInquiryResponse does not care about that anymore when it does the emails job.
I tried to make it as close to a DuckType as possible:
```
_reflections.each do |inquiry, _|
+      next unless send(:"#{inquiry}_id?")
+      ClientMailer.client(
+        response_body.transform_keys!(&:to_sym),
+        send(:"#{inquiry}")
+      ).deliver
+    end
```
The `_reflections` keyword hold all the related models.
iterating over them and sending them inquiry_id? shows me if they are truely related on creation or not.
for example if a model responds to moving_inquiry_id? that meand that on creation it was a moving_inquiry.
so we can pass the name to it trough dynamic interpolation. when it does not respond we skip it.

This approche allows me to keep the code dry. i dont expect more inquiry types to apprear soon anyways.
but should it happen i can simply add a has_one received_inquiry_response to it.
and a belongs_to xxx_inquiry, optional true. add a column to received_inquiries_responses table indexed on the key (inquiry_id)
run the migration and done.

it will work without effort. no views... no logic, nada. they all quack like ducks so lets treat them like ducks and interpolate their names.
Its super preformant.