Add module for better http_language_header extraction, fix helper too
Basically what this module does is check if :initialized is in the session.
if its not it grabs the language code from the http accept language header in the browser.
then sets that to the apps I18n.locale and also ajusts the params with the right locale /en etc.

Then it makes that :initialized session key and sets it to true.
so the next time the same session comes along... it allready has a language as default.

I added some assignements like: params[:locale] = I18n.locale in the module and in the
language switcher helpers.

Those set the currently found lang to the params.
This acually breaks the behaviour and is not very consisten.
But its good enought for our current app specs.
The language switcher works.
When user come to the page they are now welcomed in their home language.
Which is awsome! 😎 "English... do you speek it!"