Add parent thank you page url maker module for controllers, fix translations
The parent thank you page url maker module, redirects the user after form is filled and saved
to the corresponding thank you page on the parent website.

This is specific to the iframe setup. and UOS app.

Most interesting part here in this module is the:
caller_class_name = caller.first.split("/").pop.split(".rb").first

which actually finds the class name of the caller.
from there we can interpolate and redirect to the right place.
since we have multiple languages the t() method of I18n was very usefull here.