Create local thank you page, update controller and translations layer.
the local thank you page acts as a buffer page.
user get redirected here. the url is massed to it.
and then a javascript is invoked redirecting the parent.window.location.href
to the corresponding url.

It also acts as a fallback when there is no js at all (yes it is still possible in 2018)