Add TelegramNotifier after_commit on create models
When inquiries are created and commited to the database we send out a telegram notification
holding the object as json.

When responses arrive and get commited to the database notify trough telegram.

after_save is sometimes slower than the workers. after_commit is better in this case!
This resolves the ActiveJob::DeserializationError bug.

https://stackoverflow.com/questions/28561508/activejob-fails-deserializing-an-object