Add PDF assets
I kind of took a shortcut here and modified its.

I generated HTML from and existing PDF

Took it apart and choped it up and modified parts of it to work for me.

generator i used can be found here:

https://www.pdftohtml.net/

The style sheet is made with all positions being absolute!

I read the wicked_pdf does not respond well when adding too much style.
so this is a perfect solution light weight and preformant and pixel percise if you ask me.
beats the hell out of using prawn!