Refactor and dry out the Movu Transciever service object class
Added more service objects each with its own single responsibility.
HttpPostRequester: taks a json and a url as arguments and returns a response object.
HttpStatusCodeLogger: takes a status code string or integer and returns a warning
ResponseLogger: takes an Inquiry object and a response object and returns pretty details.
MovuJsonBuilder: build json out of an inquiry object specific for movu.
MovuTransiever: sends an inquiry request and recieves a response that is saved.

During this refactoring session my focus was mainly on reducing the amount of code in
the MovuTransiever, this is way cleaner now. but im not done yet.